break: adds a `ROOT_IDX` definition to `ReadNode`'s index - this allows `AllNode`s and `AnyNode`s to be implement for nodes with arbitrary indexes
break: restricts indexes allowed on `IndexDawg` to be `Copy`able and `PartialEq`
- this makes most `Dawg`s that satisfy this trait more simply implement `Dawg` methods

break: changes the `WriteNode` api to use a setter rather than a mutable reference
- this makes it simpler to implement unit conversions when setting indexes

feat: compacts both `WideNode`s and `ThinNode`s
- in theory `2^32 ~= 4e9` should be enough nodes for storing most dictionaries,
  which tend to come in at `~1e5`

+++ src/pattern/convert.rs
@@ -9,7 +9,7 @@ impl From<Pattern> for FlatDawg<ThinNode> {
 
         for mask in value.0 {
             let node = ThinNode {
-                idx: nodes.len() + repeats,
+                idx: (nodes.len() + repeats) as u32,
                 mask,
             };
             for _ in 0..repeats {
@@ -38,7 +38,7 @@ impl From<Pattern> for FlatDawg<WideNode<THIN_CHARS>> {
             let next = nodes.len() + 1;
             nodes.push(WideNode {
                 end: false,
-                children: array::from_fn(|i| (mask >> i & 1) as usize * next),
+                children: array::from_fn(|i| (mask >> i & 1) * next as u32),
             });
         }
